refactor(contexts): split SyncConfig provider from context for Fast Refresh
A file that exports a component plus a hook/context bails out of React
Fast Refresh (the '"SyncConfigContext" export is incompatible' HMR
warning, forcing full reloads). Split it: SyncConfigContext.ts holds the
context object, value type, and useSyncConfigContext hook; the
SyncConfigProvider component moves to its own SyncConfigProvider.tsx.
Also satisfies the one-component-per-file rule.

The same pattern remains in LightboxContext, SidebarLayoutContext, and
TabsContext.
