+++ src/components/App.tsx
@@ -1,5 +1,5 @@
 import { SidebarLayoutProvider } from "@/contexts/SidebarLayoutContext";
-import { SyncConfigProvider } from "@/contexts/SyncConfigContext";
+import { SyncConfigProvider } from "@/contexts/SyncConfigProvider";
 import { TabsProvider } from "@/contexts/TabsContext";
 import { useCodeThemeStyle } from "@/hooks/useCodeThemeStyle";
 import { useLocale } from "@/hooks/useLocale";
