+++ src/contexts/SyncConfigProvider.tsx
@@ -1,5 +1,6 @@
-import { createContext, type ReactNode, useContext, useMemo } from "react";
+import { type ReactNode, useMemo } from "react";
 import { useSyncConfig } from "@/hooks/useSyncConfig";
+import { SyncConfigContext, type SyncConfigContextValue } from "./SyncConfigContext";
 import { useTabsContext } from "./TabsContext";
 
 // Single source of truth for the active workspace's sync config. Both the
@@ -7,16 +8,6 @@ import { useTabsContext } from "./TabsContext";
 // disabling sync (or running one) in the modal is reflected on the pill
 // immediately — they share one `useSyncConfig` instance instead of each
 // holding a stale copy.
-
-type SyncConfigApi = ReturnType<typeof useSyncConfig>;
-
-export interface SyncConfigContextValue extends SyncConfigApi {
-  /** Active folder workspace root, or null when no folder tab is active. */
-  workspacePath: string | null;
-}
-
-export const SyncConfigContext = createContext<SyncConfigContextValue | null>(null);
-
 export function SyncConfigProvider({ children }: { children: ReactNode }) {
   const { workspace } = useTabsContext();
   const workspacePath = workspace?.root ?? null;
@@ -28,10 +19,4 @@ export function SyncConfigProvider({ children }: { children: ReactNode }) {
   );
 
   return <SyncConfigContext.Provider value={value}>{children}</SyncConfigContext.Provider>;
-}
-
-export function useSyncConfigContext(): SyncConfigContextValue {
-  const ctx = useContext(SyncConfigContext);
-  if (!ctx) throw new Error("useSyncConfigContext must be used inside <SyncConfigProvider>");
-  return ctx;
 }
